+++ SingSpaze/Models/Management/Account.cs
@@ -13,12 +13,12 @@ namespace SingSpaze.Models.Management
 
     public class Editaccount
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string username { get; set; }
         public string password { get; set; }
         public string firstname { get; set; }
         public string lastname { get; set; }
         public string email { get; set; }
-        public int facebookid { get; set; }
+        public long facebookid { get; set; }
     }
 }
