+++ SingSpaze/Models/Management/Song.cs
@@ -8,7 +8,7 @@ namespace SingSpaze.Models.Management
 {
     public class Song
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Album { get; set; }
         public string Artist { get; set; }
@@ -22,16 +22,16 @@ namespace SingSpaze.Models.Management
     public class Editsong
     {
         
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name_TH { get; set; }
         public string Name_EN { get; set; }
-        public int Artist { get; set; }
-        public int Album { get; set; }
-        public int Language { get; set; }
+        public long Artist { get; set; }
+        public long Album { get; set; }
+        public long Language { get; set; }
         public string ReleasedDate { get; set; }
         public string Length { get; set; }
-        public int Genre { get; set; }
-        public int Publisher { get; set; }
+        public long Genre { get; set; }
+        public long Publisher { get; set; }
         public string Lyrics { get; set; }
         public int Status { get; set; }
